requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
chirptext==0.1.2
Cython==3.0.11
Deprecated==1.2.14
distro==1.9.0
dyNET38==2.2
fsrs==3.1.0
fugashi==1.4.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jaconv==0.4.0
jamdict==0.1a11.post2
jamdict_data==1.5
jiter==0.6.1
lxml==5.3.0
nagisa==0.2.11
numpy==2.2.0
openai==1.52.2
plac==1.4.3
puchikarui==0.1
pydantic==2.9.2
pydantic_core==2.23.4
pykakasi==2.3.0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
termcolor==2.5.0
tqdm==4.66.5
typing_extensions==4.12.2
unidic==1.1.0
urllib3==2.3.0
wasabi==0.10.1
wrapt==1.16.0
Flask==3.0.0
Flask-CORS==4.0.0
google-cloud-texttospeech==2.24.0
google-cloud-translate==3.19.0
